# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Platform API
  description: |
    Agent Platform Product Experience API - A multi-agent platform for task orchestration 
    and execution with LLM providers, background task processing, workflows, triggers, and connectors.
    
    All endpoints use a standard API response envelope:
    - Success: `{ ok: true, data: ..., requestId: "..." }`
    - Error: `{ ok: false, error: { code: "...", message: "..." }, requestId: "..." }`
  version: "0.7.0-rc.1"
  contact:
    name: Agent Platform Team
servers:
  - url: http://localhost:3003
    description: Development
  - url: http://localhost:3103
    description: E2E Testing
...
paths:
...
  /api/v1/setup/user:
    post:
      tags: [Auth & Setup]
      summary: Create initial user
      description: Creates the first user and logs them in (only allowed when no users exist)
      operationId: createInitialUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/tool-results/{resultId}:
    get:
      tags: [Agents & Providers]
      summary: Get tool result
      description: Returns a stored tool result by ID
      operationId: getToolResult
      security:
        - sessionAuth: []
      parameters:
        - name: resultId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResultResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==========================================================================
  # Workflows
  # ==========================================================================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: agent-platform-session
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API token for programmatic access (set API_AUTH_TOKEN env var to enable)
...
  schemas:
...
    ApiEnvelope:
      type: object
      required: [ok, requestId]
      properties:
        ok:
          type: boolean
          description: true for success, false for error
        requestId:
          type: string
          description: Traceable request identifier
    ApiErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          description: Additional error details
...
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [error]
          properties:
            error:
              $ref: "#/components/schemas/ApiErrorDetail"

    # ---- Auth & Setup ----
...
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    UserMetadata:
      type: object
      required: [userId, username, createdAt]
      properties:
        userId:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [user]
              properties:
                user:
                  $ref: "#/components/schemas/UserMetadata"

    # ---- RBAC - API Keys ----
...
    ToolResultResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [resultId, toolExecutionId, toolName, userId, sizeBytes, contentType, sensitivity, createdAt]
              properties:
                resultId:
                  type: string
                toolExecutionId:
                  type: string
                toolName:
                  type: string
                userId:
                  type: string
                sessionId:
                  type: string
                sizeBytes:
                  type: integer
                contentType:
                  type: string
                preview:
                  type: string
                sensitivity:
                  type: string
                createdAt:
                  type: string
                  format: date-time

    # ---- Workflows ----
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            requestId: "abc12345"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
            requestId: "abc12345"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "CONFLICT"
              message: "Resource already exists"
            requestId: "abc12345"
...